+++ Reservation/Domain/予約/ReserverId.cs
@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Reservation.Domain {
+namespace Reservation.Domain.予約 {
     // 予約者ID
     public class ReserverId {
