+++ Reservation/Domain/予約/会議室/MeetingRoomName.cs
@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Reservation.Domain 
+namespace Reservation.Domain.—\–ñ.‰ï‹cŽº 
 {
     public enum MeetingRoomName 
     {
