+++ Reservation/Domain/予約/想定使用人数.cs
@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Reservation.Domain {
+namespace Reservation.Domain.予約 {
     public class 想定使用人数 {
