+++ Reservation/Domain/予約/期間/予約年月日.cs
@@ -1,9 +1,7 @@
 using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Text;
 
-namespace Reservation.Domain {
+namespace Reservation.Domain.予約.期間 {
     /// <summary>
     /// 年月日だけ持つ
     /// </summary>
